Add files via upload
This is a starting point for my instant animation project, containing any related files.  Final goal of the  project is to have an interface  in Blender where one can just click a button for a particular type of character, and the character bones  will appear, they can then attach their mesh to the bones, and  hit the play button and the character will walk, run, trot, gallop, etc. with no need for IK.